Add files via upload
Code is last operational version, circa 1994. It is roughly Fortran 66, and will not compile straightforwardly on modern cmopilers. The skiles1.f is more likely.  (This name comes from the fact that the initial implementation by Ed O'Lenic in ca. 1978 was based on a paper by Frank Skiles.) Dust off your Hollerith.